Fix DB problem
If you first created a study and then performed a study update, the
updated data would not be persistently saved into the database, as
the database was no longer in autocommit mode after the study was
inserted.  As the study insert code is working fine, I removed the
transaction system from the study insert.